requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.2
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lightning-utilities==0.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.0b1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
yarl==1.9.4